pyproject.toml
[tool.poetry]
name = "elrond"
version = "0.1.0"
description = ""
authors = ["Reyhanishere <reyhaneh.parviziyan@gmail.com>"]
readme = "README.md"
packages = [
    { include = "elrond" }
]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
django = "^4.2"  # Downgrade Django to match django-allauth
"django-allauth" = { extras = ["socialaccount"], version = "^0.58.2" }
psycopg = { extras = ["binary"], version = "^3.2.9" }
"dj-database-url" = "^2.3.0"
"django-browser-reload" = "^1.18.0"
"django-widget-tweaks" = "^1.5.0"
python-decouple = "^3.8"
python-dotenv = "^1.1.0"

[tool.poetry.group.dev.dependencies]
django-debug-toolbar = "^5.2.0"

[build-system]
requires = ["poetry-core>=2.0.0"]
build-backend = "poetry.core.masonry.api"